Actualizando repo

+++ init.py
@@ -3,6 +3,8 @@
 github_repo = 'repos-especializacion-UdeA/monografia_modelos'
 zip_file_url="https://github.com/%s/archive/master.zip"%github_repo
 
+import requests, zipfile, io, os, shutil
+
 def get_last_modif_date(localdir):
     try:
         import time, os, pytz
@@ -14,18 +16,25 @@ def get_last_modif_date(localdir):
     except Exception:
         return None
     
-import requests, zipfile, io, os, shutil
 def init(force_download=False):
     if force_download or not os.path.exists("local"):
         print("replicating local resources")
-        dirname = "-master/"
+        dirname = github_repo.split("/")[-1]+"-main/"
         if os.path.exists(dirname):
             shutil.rmtree(dirname)
         r = requests.get(zip_file_url)
+        print("Downloading from %s"%zip_file_url)
         z = zipfile.ZipFile(io.BytesIO(r.content))
         z.extractall()
         if os.path.exists("local"):
             shutil.rmtree("local")
-        shutil.move(dirname+"/local", "local")
+        if os.path.exists("local"):
+            shutil.rmtree("local")
+        if os.path.exists(dirname+"/content/local"):
+            shutil.move(dirname+"/content/local", "local")
+        elif os.path.exists(dirname+"/local"):
+            shutil.move(dirname+"/local", "local")
         shutil.rmtree(dirname)
+        print("ok")
+
